Projects components renamed to singular project
