+++ src/components/ProjectGallery.vue
