+++ src/components/ProjectInfo.vue
