+++ src/views/SingleProject.vue
@@ -1,33 +1,33 @@
 <template>
 	<div class="container mx-auto mt-10 sm:mt-20">
 		<!-- Project header -->
-		<ProjectSingleHeader :singleProjectHeader="singleProjectHeader" />
+		<ProjectHeader :singleProjectHeader="singleProjectHeader" />
 
 		<!-- Project gallery -->
-		<ProjectSingleGallery :projectImages="projectImages" />
+		<ProjectGallery :projectImages="projectImages" />
 
 		<!-- Project information -->
-		<ProjectSingleInfo :projectInfo="projectInfo" />
+		<ProjectInfo :projectInfo="projectInfo" />
 
 		<!-- Project related projects -->
-		<ProjectSingleRelatedProjects :relatedProject="relatedProject" />
+		<ProjectRelatedProjects :relatedProject="relatedProject" />
 	</div>
 </template>
 
 <script>
 import feather from 'feather-icons';
-import ProjectSingleHeader from '../components/ProjectSingleHeader.vue';
-import ProjectSingleGallery from '../components/ProjectSingleGallery.vue';
-import ProjectSingleInfo from '../components/ProjectSingleInfo.vue';
-import ProjectSingleRelatedProjects from '../components/ProjectSingleRelatedProjects.vue';
+import ProjectHeader from '../components/ProjectHeader.vue';
+import ProjectGallery from '../components/ProjectGallery.vue';
+import ProjectInfo from '../components/ProjectInfo.vue';
+import ProjectRelatedProjects from '../components/ProjectRelatedProjects.vue';
 
 export default {
 	name: 'Projects',
 	components: {
-		ProjectSingleHeader,
-		ProjectSingleGallery,
-		ProjectSingleInfo,
-		ProjectSingleRelatedProjects,
+		ProjectHeader,
+		ProjectGallery,
+		ProjectInfo,
+		ProjectRelatedProjects,
 	},
 	data: () => {
 		return {
